+++ sickbeard/scene_exceptions.py
@@ -194,7 +194,7 @@ def get_scene_exception_by_name_multiple(show_name):
     :rtype: Tuple[None, None, None] or Tuple[int, int or long, int]
     """
     try:
-        exception_result = name_cache.nameCache[helpers.full_sanitize_scene_name(show_name)]
+        exception_result = name_cache.sceneNameCache[helpers.full_sanitize_scene_name(show_name)]
         return [exception_result]
     except (BaseException, Exception):
         return [[None, None, None]]
@@ -294,6 +294,7 @@ def retrieve_exceptions():
 
     if cl:
         my_db.mass_action(cl)
+        name_cache.buildNameCache(update_only_scene=True)
 
     # since this could invalidate the results of the cache we clear it out after updating
     if changed_exceptions:
@@ -338,6 +339,8 @@ def update_scene_exceptions(tvid, prodid, scene_exceptions):
         my_db.action('INSERT INTO scene_exceptions'
                      ' (indexer, indexer_id, show_name, season) VALUES (?,?,?,?)',
                      [tvid, prodid, cur_exception, cur_season])
+
+    sickbeard.name_cache.buildNameCache(update_only_scene=True)
 
 
 def _anidb_exceptions_fetcher():